Beta refactoring -> BetaTag
