+++ src/components/Option/Settings/general-settings.tsx
@@ -7,7 +7,7 @@ import { TTSModeSettings } from "./tts-mode"
 import { useStorage } from "@plasmohq/storage/hook"
 import { SystemSettings } from "./system-settings"
 import { SSTSettings } from "./sst-settings"
-import { BetaTag } from "@/components/Common/Beta"
+import { BetaTag } from "@/components/Common/BetaTag"
 import { getDefaultOcrLanguage, ocrLanguages } from "@/data/ocr-language"
 import { Storage } from "@plasmohq/storage"
 import { useQuery } from "@tanstack/react-query"
